books-catalogue example with codegen
