+++ examples/books-catalogue-codegen/src/main/java/dev/springharvest/library/codegen/BooksCatalogueCodegenApplication.java
@@ -1,0 +1,13 @@
+package dev.springharvest.library.codegen;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BooksCatalogueCodegenApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(BooksCatalogueCodegenApplication.class, args);
+  }
+
+}
