+++ examples/books-catalogue-codegen/src/main/resources/.env
@@ -1,0 +1,8 @@
+DB_USERNAME=changeme
+DB_PASSWORD=changeme
+DB_HOST=localhost
+DB_PORT=5434
+DB_NAME=generated-library-db
+SPRING_DATASOURCE_DRIVER-CLASS-NAME=org.postgresql.Driver
+SPRING_DATASOURCE_USERNAME=${DB_USERNAME}
+SPRING_DATASOURCE_PASSWORD=${DB_PASSWORD}
