+++ examples/books-catalogue-codegen/src/main/resources/application-liquibase.properties
@@ -1,3 +1,4 @@
 logging.level.liquibase=DEBUG
-spring.liquibase.change-log=classpath:db.changelog/db.changelog-master.yaml
+spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.yaml
 spring.liquibase.drop-first=true
+spring.liquibase.enabled=true
