+++ examples/books-catalogue-codegen/src/main/resources/application-postgres.properties
@@ -1,11 +1,11 @@
 DB_USERNAME=changeme
 DB_PASSWORD=changeme
 DB_HOST=localhost
-DB_PORT=5433
-DB_NAME=library
+DB_PORT=5434
+DB_NAME=generated-library
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.username=${DB_USERNAME}
 spring.datasource.password=${DB_PASSWORD}
 spring.datasource.url=jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}
-#spring.docker.compose.enabled=false
-#spring.docker.compose.file=docker-compose.yaml
+spring.docker.compose.enabled=true
+spring.docker.compose.file=docker-compose.yaml
