+++ examples/books-catalogue-codegen/src/main/resources/application-test.properties
@@ -1,0 +1,14 @@
+#logging.level.root=DEBUG
+#logging.level.web=DEBUG
+server.servlet.context-path=/api
+## LOAD ENV VARIABLES
+spring.config.import=optional:.env[.properties]
+## DATABASE
+spring.docker.compose.enabled=false
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:tc:postgresql:15.2-bullseye:///spring_boot_testcontainers
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.sql.init.platform=postgres
+spring.liquibase.contexts=test
+spring.liquibase.change-log=db/changelog/db.changelog-master.yaml
