+++ examples/books-catalogue-codegen/src/main/resources/application.properties
@@ -1,17 +1,31 @@
-server.servlet.context-path=/api
-logging.level.root=ERROR
-# API Documentation
-springdoc.api-docs.path=/docs
-springdoc.swagger-ui.path=/docs/ui
-springdoc.swagger-ui.tagsSorter=alpha
-springdoc.swagger-ui.operationsSorter=alpha
-# Database
+# LOGGING
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
-# TODO: Identify if transaction batch size is sufficient for import
+# CLI
+#spring.banner.location=classpath:banner.txt
+spring.output.ansi.enabled=always
+# CONFIG
+server.servlet.context-path=/api
+## TODO remove profiles for prod
+spring.profiles.default=dev
+## BATCH
 # https://www.baeldung.com/jpa-hibernate-batch-insert-update
 spring.jpa.properties.hibernate.jdbc.batch_size=50
 spring.servlet.multipart.max-file-size=50MB
 spring.servlet.multipart.max-request-size=50MB
 spring.main.allow-circular-references=true
-spring.threads.virtual.enabled=true
+## API Documentation
+springdoc.api-docs.path=/docs
+springdoc.swagger-ui.path=/docs/ui
+springdoc.swagger-ui.tagsSorter=alpha
+springdoc.swagger-ui.operationsSorter=alpha
+## LOAD ENV VARIABLES
+spring.config.import=optional:.env[.properties]
+## DATABASE
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
+spring.datasource.url=jdbc:postgresql://${DB_HOST}:${DB_PORT}/${DB_NAME}
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+# Enable PostGIS extension
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
