--- examples/books-catalogue-codegen/src/main/resources/db.changelog/BASE_PARENTS/DDL/create_base_trace_parent_table.sql
@@ -1,11 +1,0 @@
---liquibase formatted sql
---changeset BillyBolton:create-base-trace-parent-table
-
-DROP TABLE IF EXISTS base_trace_parent CASCADE;
-CREATE TABLE base_trace_parent
-(
-    created_by UUID,
-    date_created TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
-    updated_by UUID,
-    date_updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
-);
