--- examples/books-catalogue-codegen/src/main/resources/db.changelog/EXTENSIONS/DDL/create_db_extensions.sql
@@ -1,4 +1,0 @@
---liquibase formatted sql
---changeset BillyBolton:create-db-extensions
-
-CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
