+++ examples/books-catalogue-codegen/src/main/resources/db/changelog/AUTHORS/DDL/create_authors_table.sql
