+++ examples/books-catalogue-codegen/src/main/resources/db/changelog/BOOKS/DDL/create_books_table.sql
