+++ examples/books-catalogue-codegen/src/main/resources/db/changelog/BOOKS/DML/insert_data_books.sql
