+++ examples/books-catalogue-codegen/src/main/resources/db/changelog/PUBLISHERS/DDL/create_publishers_table.sql
