+++ examples/books-catalogue-codegen/src/main/resources/db/changelog/PUBLISHERS/DML/insert_data_publishers.sql
