--- examples/books-catalogue-codegen/src/test/java/dev/springharvest/library/BooksCatalogueCodegenApplicationTests.java
@@ -1,13 +1,0 @@
-package dev.springharvest.library;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BooksCatalogueCodegenApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
