+++ examples/books-catalogue-codegen/src/test/java/dev/springharvest/library/codegen/BooksCatalogueCodegenApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.springharvest.library.codegen;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BooksCatalogueCodegenApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
