+++ examples/books-catalogue-codegen/src/test/java/dev/springharvest/library/codegen/config/MockThirdPartyApiConfig.java
@@ -1,0 +1,10 @@
+package dev.springharvest.library.codegen.config;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.mock.mockito.MockBean;
+
+@TestConfiguration
+@MockBean({})
+public class MockThirdPartyApiConfig {
+
+}
