+++ examples/books-catalogue-codegen/src/test/java/dev/springharvest/library/codegen/config/TestComponentScanningConfig.java
@@ -1,0 +1,10 @@
+package dev.springharvest.library.codegen.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = {"dev.springharvest.testing"})
+public class TestComponentScanningConfig {
+
+}
