--- examples/books-catalogue-codegen/src/test/java/dev/springharvest/library/config/TestContainerConfig.java
@@ -1,34 +1,0 @@
-package dev.springharvest.library.config;
-
-import javax.sql.DataSource;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.context.annotation.Primary;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.testcontainers.containers.PostgreSQLContainer;
-
-@TestConfiguration
-@Lazy
-public class TestContainerConfig {
-
-  @Bean(initMethod = "start", destroyMethod = "stop")
-  public PostgreSQLContainer<?> postgreSQLContainer() {
-    PostgreSQLContainer<?> container = new PostgreSQLContainer<>("postgres:13").withDatabaseName("test")
-        .withUsername("test")
-        .withPassword("test");
-    return container;
-  }
-
-  @Bean
-  @Primary
-  public DataSource dataSource(PostgreSQLContainer<?> container) {
-    DriverManagerDataSource dataSource = new DriverManagerDataSource();
-    dataSource.setDriverClassName("org.postgresql.Driver");
-    dataSource.setUrl(container.getJdbcUrl());
-    dataSource.setUsername(container.getUsername());
-    dataSource.setPassword(container.getPassword());
-    return dataSource;
-  }
-
-}
